Add rubocop-rspec_rails and get rubocop run to green
Rubcop wouldn't even run on master because the latest rubocop-rspec has
extracted some cops to separate sub gems and would fail on parsing the todo
file which referenced one of those extracted cops. Introducing the missing gem,
rubocop-rspec_rails and changing the name in the config gets us to running
rubocop. There is one infraction that exists in the source so we fix that to
get us to a green rubocop run.